requirements.txt
django
djangorestframework
pymongo
mongoengine
streamlit
python-dotenv